feat: Add endpoint to get trainers
This commit adds a new endpoint `/trainers` to retrieve a list of trainers. The `GetTrainers` function in the `handlers` package is implemented to fetch trainers from the database and return them as a JSON response. This feature enhances the functionality of the application by allowing clients to view and select trainers for their training needs.