feat: Add CreatedByID field to TrainingPlan model
This commit adds a new field `CreatedByID` to the `TrainingPlan` model. This field is used to store the ID of the user who created the training plan. The `CreatedByID` field is added to the JSON representation of the `TrainingPlan` model as well. This feature enhances the functionality of the application by allowing the tracking of the user who created each training plan.