```text feat: Add helper function to generate avatar URLs for users
This commit adds a new helper function `GenerateAvatarURL` to the `helpers` package. The function takes an email and role as input and generates a unique avatar URL based on the email. The role parameter determines the avatar set to be used. This feature enhances the visual representation of users in the application by providing personalized avatar images.