feat: Add 'unit' field to Exercise model
This code change adds a new 'unit' field to the Exercise model in the exercise.go file. The 'unit' field represents the unit of measurement for the exercise, which can be either "time" or "duration". This feature enhances the flexibility of the application by allowing exercises to be categorized based on their unit of measurement.