refactor: Update ClientWorkout model to include DaysUntilNext
This code change in the `models/client_workout.go` file updates the `ClientWorkout` model to include a new field `DaysUntilNext`. This field represents the number of days until the next workout for the client. The `DaysUntilNext` field is added after the `Description` field and before the `Order` field. This refactor enhances the functionality of the application by providing a way to track and display the time until the next workout for clients.