refactor: Update LoginUser function to preload TrainerProfile.Specialties
This code change in the `handlers/user.go` file updates the `LoginUser` function to preload the `TrainerProfile.Specialties` field when querying the user from the database. This refactor enhances the functionality of the application by including the specialties of the trainer in the response, providing more comprehensive information about the user during the login process.