refactor: Add endpoint to get clients by trainer ID
This code change adds a new endpoint `/trainers/:id/clients` in the `trainer_routes.go` file. The endpoint calls the `GetClientsByTrainerID` function in the `trainer.go` file to retrieve the clients associated with a specific trainer ID. This refactor enhances the functionality of the application by providing a way to fetch clients based on their trainer.