refactor: Add GetUserTrainingPlans endpoint
This code change adds a new endpoint `/users/:id/training_plans` in the `user_routes.go` file. The endpoint calls the `GetUserTrainingPlans` function in the `user.go` file to retrieve the training plans associated with a specific user ID. This refactor enhances the functionality of the application by providing a way to fetch training plans for a user.